Added profilePic to user table
Register has the block of code to add pictures but I can't get it to work so I commented it out; will need someone to help me debug it tmr or later.
Profile handlebars has been changed to have a profile picture to display, a reviews box and a order history box. Will need to integrate this with James and Jiang Rui later.  Added a /user/upload for the users to be able to upload any images on their end. (maybe its because I didnt have a /upload for guests to upload their images.. shit.) Will continue after my lunch.